Rename lib to have lib in the name and work w/ node
